llm_exp/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
openai==1.54.0
pydantic==2.5.3
python-dotenv==1.0.0
